resample completed
