+++ egrin2/resample.py
@@ -1,191 +1,138 @@
-def rsd( self, vals ):
+#!/usr/bin/env python
+
+"""Resample ratios. Add them to col_resample collection in an existing egrin2 MongoDB"""
+
+__author__ = "Aaron Brooks"
+__copyright__ = "Copyright 2014, cMonkey2"
+__credits__ = ["Aaron Brooks"]
+__license__ = "GPL"
+__version__ = "1.0.1"
+__maintainer__ = "Aaron Brooks"
+__email__ = "brooksan@uw.edu"
+__status__ = "Development"
+
+import random
+from multiprocessing import Pool
+
+# $ hg clone ssh://hg@bitbucket.org/djcbeach/monary ./monary
+# $ cd ./monary && python setup.py install
+from monary import Monary
+
+from pymongo import MongoClient
+import numpy as np
+import pandas as pd
+from joblib import Parallel, delayed
+
+def multiSample(number):
+	"""Returns how many documents in the collection"""
+	colResampleInd( rows = rows, cols = [server_number])
+ 
+def rsd( vals ):
 	return abs( np.std( vals ) / np.mean( vals ) )
 
-def colResampleInd( self, n_rows, col, n_resamples, keepP ):
+def resample( row_vals, n_rows ):
+		return rsd( random.sample( row_vals, n_rows ) )
+
+def choose_n( col, vals, n, add, client, db, n_rows, n_resamples ):
+	normalized = vals.loc[:,"normalized_expression"].copy()
+	normalized.sort()
+	standardized = vals.loc[:,"standardized_expression"].copy()
+	standardized.sort()
+
+	if add:
+		d = {
+			"n_rows": n_rows,
+			"col_id": col,
+			"resamples": n_resamples,
+			"lowest_normalized": normalized.iloc[0:n].tolist(),
+	 		"lowest_standardized": standardized.iloc[0:n].tolist()
+		}
+		client[ db ][ "col_resample" ].insert( d )
+	else:
+		#update
+		resamples = n_resamples + old_records[ col ][ "resamples" ]
+		n2keep = round( resamples*keepP )
+		ran = normalized + old_records[ col ][ "lowest_normalized" ]
+		ran.sort()
+		ran = ran.iloc[ 0: int( n2keep ) ].tolist()
+		ras = standardized + old_records[ col ][ "lowest_standardized" ]
+		ras.sort()
+		ras = ras.iloc[ 0: int( n2keep ) ].tolist()
+		client[ db ][ "col_resample" ].update( { "n_rows": n_rows, "col_id": col }, { "$set": { "resamples": resamples, "lowest_normalized": ran, "lowest_standardized": ras } } )
+
+def colResampleInd( host, n_rows, cols, n_resamples = 20000, keepP = 0.1, port = 27017, db = "egrin2_db" ):
 	"""Resample gene expression for a given number of genes in a particular condition using RSD"""
 
-	def resample( n_rows, row_vals ):
-		val_rsd = self.rsd( random.sample( row_vals, n_rows ) )
-		
+	print n_rows
 
-	n2keep = round( n_resamples*keepP )
-
-	# get ratios for col 
-	row_vals_normalized = [ i[ "normalized_expression" ] for i in self.db.gene_expression.find( { "col_id":col } ) if isinstance( i[ "normalized_expression" ], float) ]
-	row_vals_standardized = [ i[ "standardized_expression" ] for i in self.db.gene_expression.find( { "col_id":col } ) if isinstance( i[ "standardized_expression" ], float) ]
-
-	resample_array_normalized = [ self.rsd( random.sample( row_vals_normalized, n_rows ) ) for i in range( 0, n_resamples ) ]
-	resample_array_normalized.sort( )
-	resample_array_normalized = resample_array_normalized[ 0 : int( n2keep ) ]
-	
-	resample_array_standardized = [ self.rsd( random.sample( row_vals_standardized, n_rows ) ) for i in range( 0, n_resamples ) ]
-	resample_array_standardized.sort( )
-	resample_array_standardized = resample_array_standardized[ 0 : int( n2keep ) ]
+	# make connection
+	client = MongoClient( host = host, port=port )
 
 	# see if a record for this gene set size exists 
-	old_record = self.db.col_resample.find_one( { "n_rows": n_rows, "col_id": col } )
-	if  old_record is not None:
-		# entry already exists for this. add new resamples to it. 
-		resamples = n_resamples + old_record[ "resamples" ]
-		n2keep = round( resamples*keepP )
-		ran = resample_array_normalized + old_record[ "lowest_normalized" ]
-		ran.sort()
-		ran = ran[ 0: int( n2keep ) ]
-		ras = resample_array_standardized + old_record[ "lowest_standardized" ]
-		ras.sort()
-		ras = ras[ 0: int( n2keep ) ]
-		self.db.col_resample.update( { "n_rows": n_rows, "col_id": col }, { "$set": { "resamples": resamples, "lowest_normalized": ran, "lowest_standardized": ras } } )
-	else:
-		d = {
-		"n_rows": n_rows,
-		"col_id": col,
-		"resamples": n_resamples,
-		"lowest_normalized": resample_array_normalized,
- 			"lowest_standardized": resample_array_standardized
-		}
-		self.db.col_resample.insert( d )
+	old_records = { i[ "col_id" ] : i for i in client[ db ][ "col_resample" ].find( { "n_rows": n_rows, "col_id": { "$in": cols } } ) }
+
+	if  old_records is not None:
+		toUpdate = [ int( i[ "col_id" ] ) for i in old_records.values() if int( i[ "resamples" ] ) < n_resamples ]
+		
+	toAdd = [ i for i in cols if i not in old_records.keys() ]
+
+	if len( toAdd ) == 0 and len( toUpdate ) == 0:
+		print "Nothing to add"
+		client.close()
+		return None
+
+	n2keep = int( round( n_resamples*keepP ) )
+
+	# toAdd
+	if len( toAdd ) > 0:
+		print "Adding new document to MongoDB"
+		with Monary( 'mongodb://'+host+':'+str(port)+'/' ) as mon:
+			mon.connect()
+			columns = ["col_id", "normalized_expression", "standardized_expression"]
+			arrays = mon.query(
+				"egrin2_db",                         # database name
+				"gene_expression",                   # collection name
+				{ "col_id": { "$in": toAdd } },                             # query spec
+				columns, # field names (in Mongo record)
+				["int64", "float64", "float64"] 
+			)
+			df = np.matrix(arrays).transpose()
+			df = pd.DataFrame(df, columns=columns)
+
+		df = df.groupby("col_id")
+		df_rsd = pd.concat( [ df.aggregate( resample, n_rows ) for i in range( 0, n_resamples ) ] )
+		df_rsd = df_rsd.groupby( df_rsd.index )
+		tmp = [ choose_n( int( i ), df_rsd.get_group( i ), n2keep, True, client, db, n_rows, n_resamples ) for i in df_rsd.groups.keys() ]
+
+	# toUpdate
+	if len( toUpdate ) > 0:
+		print "Updating existing MongoDB documents"
+		with Monary( 'mongodb://'+host+':'+str(port)+'/' ) as mon:
+			mon.connect()
+			columns = ["col_id", "normalized_expression", "standardized_expression"]
+			arrays = mon.query(
+				"egrin2_db",                         # database name
+				"gene_expression",                   # collection name
+				{ "col_id": { "$in": toUpdate } },                             # query spec
+				columns, # field names (in Mongo record)
+				["int64", "float64", "float64"] 
+			)
+			df = np.matrix(arrays).transpose()
+			df = pd.DataFrame(df, columns=columns)
+
+		df = df.groupby("col_id")
+		resamples = n_resamples - np.min( [ i[ "resamples" ] for i in old_records.values( ) ] )
+		if resamples > 0:
+			df_rsd = pd.concat( [ df.aggregate( resample, n_rows ) for i in range( 0, resamples ) ] )
+			df_rsd = df_rsd.groupby( df_rsd.index )
+			tmp = [ choose_n( int( i ), df_rsd.get_group( i ), n2keep, False, client, db, n_rows, resamples ) for i in df_rsd.groups.keys() ]
+	client.close()
+
 	return None
 
-def checkDBresamples( self, col, n_rows, n_resamples ):
-		if self.db.col_resample.find_one( { "n_rows": n_rows, "col_id": col, "resamples": { "$gte": n_resamples } } ) is None:
-			return col
-
-def checkRow( self, row ):
-	"""Check name format of rows. If necessary, translate."""
-	if row in self.row2id.keys():
-		return self.row2id[ row ] 
-	elif int( row ) in self.id2row.keys():
-		return int( row )
-	else:
-		print "ERROR: Cannot identify row name: %s" % row
-		return None
-
-def checkCol( self, col ):
-	"""Check name format of rows. If necessary, translate."""
-	if col in self.col2id.keys():
-		return self.col2id[ col ] 
-	elif int( col ) in self.id2col.keys():
-		return int( col )
-	else:
-		print "ERROR: Cannot identify col name: %s" % col
-		return None
-
-def rowsColPval( self, rows = None, cols = None, standardized = None, sig_cutoff = None ):
-
-	def empirical_pval( i, random_rsd, resamples ):
-		for x in range( 0, len( i ) ):
-			val = ( float ( sum( [ i.values[ 0 ] >= y for y in random_rsd[ i.index[0] ] ] ) ) / len( random_rsd[ i.index[0] ] ) ) * ( float( len( random_rsd[ i.index[0] ] ) ) / resamples[ i.index[0] ] )
-			if val >= ( float( len( random_rsd[ i.index[0] ] ) ) / resamples[ i.index[0] ] ):
-				#return ">= %f" % ( round( float( len( random_rsd[ "lowest_normalized" ] ) ) / random_rsd[ "resamples" ], 2 ) )
-				return round( float( len( random_rsd[ i.index[0] ] ) ) / resamples[ i.index[0] ], 2 )
-			elif val == 0:
-				#return "< %f" % ( 1 / float( random_rsd[ "resamples" ] ) )
-				return ( 1 / float( resamples[ i.index[0] ] ) )
-			else:
-				return val
-
-	rows = [ self.checkRow( i ) for i in rows ]
-	rows = [ i for i in rows if i is not None]
-	if rows is None:
-		print "Please provide an appropriately named array of rows"
-		return None 
-	
-	cols = [ self.checkCol( i ) for i in cols ]
-	cols = [ i for i in cols if i is not None]
-	if cols is None:
-		print "Please provide an appropriately named array of cols"
-		return None
-
-	if standardized is None:
-		# compute RSD on standardized gene expression by default
-		# other option 'False' for normalized (not standardized expression)
-		standardized = True
-
-	if sig_cutoff is None:
-		# return only cols equal below sig_cutoff
-		sig_cutoff = 0.05
-
-	exp_df = pd.DataFrame( )
-
-	# compute actual RSD
-	if standardized:
-		 exp_df =  exp_df.append( [ pd.Series( [ i[ "col_id" ], i[ "standardized_expression" ] ], index = [ "col_id", "value" ] ) for i in self.db.gene_expression.find( { "col_id": { "$in" : cols }, "row_id": { "$in" : rows } } ) if isinstance( i[ "standardized_expression" ], float ) ], ignore_index = True )
-		 # get random resamples for this col
-		 random_rsd = { i[ "col_id" ]: i[ "lowest_standardized" ] for i in self.db.col_resample.find( { "n_rows": len( rows ), "col_id": { "$in" : cols } } ) }
-		 resamples = { i[ "col_id" ]: i[ "resamples" ] for i in self.db.col_resample.find( { "n_rows": len( rows ), "col_id": { "$in" : cols } } ) }
-	else:
-		exp_df =  exp_df.append( [ pd.Series( [ i[ "col_id" ], i[ "normalized_expression" ] ], index = [ "col_id", "value" ] ) for i in self.db.gene_expression.find( { "col_id": { "$in" : cols }, "row_id": { "$in" : rows } } ) if isinstance( i[ "normalized_expression" ], float ) ], ignore_index = True )
-		# get random resamples for this col
-		random_rsd = { i[ "col_id" ]: i[ "lowest_normalized" ] for i in self.db.col_resample.find( { "n_rows": len( rows ), "col_id": { "$in" : cols } } ) }
-		resamples = { i[ "col_id" ]: i[ "resamples" ] for i in self.db.col_resample.find( { "n_rows": len( rows ), "col_id": { "$in" : cols } } ) }
-
-	if random_rsd is None:
-		print "Could not find resample DB entry for %i rows" % ( len(rows) )
-		return None
-
-	exp_df_rsd = exp_df.groupby("col_id").aggregate(self.rsd)
-
-	pvals = exp_df_rsd.groupby(level=0).aggregate(empirical_pval, random_rsd, resamples )
-	pvals.columns = ["pval"]
-	pvals.index = [ self.id2col[ i ] for i in pvals.index.values]
-
-	if sig_cutoff is not None:
-		pvals = pvals[ pvals.loc[ :,"pval" ] <= sig_cutoff ]
-
-	return pvals
-	
-def colResampleGroup( self, rows = None, cols = None, n_resamples = None, sig_cutoff = None, standardized = None ):
-	"""Resample gene expression for a given set of genes in any number of conditions. Should be used instead of colReampleInd (it calls that function)"""
-
-	if n_resamples is None:
-		n_resamples = 20000
-
-	if standardized is None:
-		# compute RSD on standardized gene expression by default
-		# other option 'False' for normalized (not standardized expression)
-		standardized = True
-
-	if sig_cutoff is None:
-		# return only cols equal below sig_cutoff
-		sig_cutoff = 0.05
-
-	rows = [ self.checkRow( i ) for i in rows ]
-	rows = [ i for i in rows if i is not None]
-	if rows is None:
-		print "Please provide an appropriately named array of rows"
-		return None 
-	
-	cols = [ self.checkCol( i ) for i in cols ]
-	cols = [ i for i in cols if i is not None]
-	if cols is None:
-		print "Please provide an appropriately named array of cols"
-		return None
-	
-	# Determine what/how many resamples need to be added to db
-	toAdd = [ self.checkDBresamples( i, len( rows ), n_resamples ) for i in cols]
-	toAdd = [ i for i in toAdd if i is not None]
-
-	count = 1
-	if len( toAdd) > 0:
-		print "I need to perform %i random resample(s) of size %i to compute pvals. Please be patient. This may take a while..." % ( len(toAdd), n_resamples )
-		for i in toAdd:		
-			currentEntry = self.db.col_resample.find_one( { "n_rows": len( rows ), "col_id": i } )
-			if currentEntry is not None:
-				# only add enough resamples to reach requested value
-				self.colResampleInd( len( rows ), i, n_resamples - currentEntry[ "resamples" ], .1 )
-			else:
-				self.colResampleInd( len( rows ), i, n_resamples, .1 )
-			if round( ( float( count) / len(toAdd) ) * 100, 2 ) % 10  == 0:
-				print "%d percent" % ( round( ( float( count ) / len( toAdd ) ) * 100, 2 ) )
-			count = count + 1
-		print "Done adding random resamples."
-
-	print "Calculating pvals"
-
-	pvals = self.rowsColPval( rows, cols, standardized, sig_cutoff )
-
-	return pvals
-
 if __name__ == '__main__':
-    cl = Calculate()
-    print cl.run()
+	cols = range(0,400)
+	tmp = Parallel(n_jobs=4)( delayed( colResampleInd )( "localhost", i, cols, n_resamples = 10) for i in range( 3,10 ) )
+
+	colResampleInd( host, n_rows, cols, n_resamples = 20000, keepP = 0.1, port = 27017, db = "egrin2_db" )
+
